move descriptor pkg to common
