+++ client/covenant/cli.go
@@ -9,8 +9,8 @@ import (
 	"github.com/ark-network/ark/client/interfaces"
 	"github.com/ark-network/ark/client/utils"
 	"github.com/ark-network/ark/common"
+	"github.com/ark-network/ark/common/descriptor"
 	"github.com/ark-network/ark/common/tree"
-	"github.com/ark-network/ark/pkg/descriptor"
 	"github.com/decred/dcrd/dcrec/secp256k1/v4"
 	"github.com/urfave/cli/v2"
 	"github.com/vulpemventures/go-elements/address"
