+++ common/descriptor/ark.go
