+++ common/descriptor/parser.go
