+++ common/descriptor/types.go
