+++ go.work
@@ -5,7 +5,6 @@ use (
 	./client
 	./common
 	./pkg/client-sdk
-	./pkg/descriptor
 	./server
 	./server/pkg/kvdb
 	./server/pkg/macaroons
