--- pkg/descriptor/go.mod
@@ -1,27 +1,0 @@
-module github.com/ark-network/ark/pkg/descriptor
-
-go 1.22.6
-
-require (
-	github.com/ark-network/ark/common v0.0.0-20240902212412-b5e1ea4d574b
-	github.com/btcsuite/btcd v0.24.2
-	github.com/btcsuite/btcd/btcec/v2 v2.3.4
-	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.3.0
-	github.com/stretchr/testify v1.9.0
-)
-
-require (
-	github.com/btcsuite/btcd/btcutil v1.1.5 // indirect
-	github.com/btcsuite/btcd/btcutil/psbt v1.1.9 // indirect
-	github.com/btcsuite/btcd/chaincfg/chainhash v1.1.0 // indirect
-	github.com/btcsuite/btclog v0.0.0-20170628155309-84c8d2346e9f // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/decred/dcrd/crypto/blake256 v1.0.1 // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	github.com/vulpemventures/fastsha256 v0.0.0-20160815193821-637e65642941 // indirect
-	github.com/vulpemventures/go-elements v0.5.4 // indirect
-	golang.org/x/crypto v0.26.0 // indirect
-	golang.org/x/sys v0.24.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
